+++ harfbuzz/src/hb-common.cc
@@ -133,65 +133,6 @@ hb_tag_to_string (hb_tag_t tag, char *buf)
   buf[3] = (char) (uint8_t) (tag >>  0);
 }
 
-
-/* hb_direction_t */
-
-const char direction_strings[][4] = {
-  "ltr",
-  "rtl",
-  "ttb",
-  "btt"
-};
-
-/**
- * hb_direction_from_string:
- * @str: (array length=len) (element-type uint8_t):
- * @len:
- *
- *
- *
- * Return value:
- *
- * Since: 0.9.2
- **/
-hb_direction_t
-hb_direction_from_string (const char *str, int len)
-{
-  if (unlikely (!str || !len || !*str))
-    return HB_DIRECTION_INVALID;
-
-  /* Lets match loosely: just match the first letter, such that
-   * all of "ltr", "left-to-right", etc work!
-   */
-  char c = TOLOWER (str[0]);
-  for (unsigned int i = 0; i < ARRAY_LENGTH (direction_strings); i++)
-    if (c == direction_strings[i][0])
-      return (hb_direction_t) (HB_DIRECTION_LTR + i);
-
-  return HB_DIRECTION_INVALID;
-}
-
-/**
- * hb_direction_to_string:
- * @direction:
- *
- *
- *
- * Return value: (transfer none):
- *
- * Since: 0.9.2
- **/
-const char *
-hb_direction_to_string (hb_direction_t direction)
-{
-  if (likely ((unsigned int) (direction - HB_DIRECTION_LTR)
-              < ARRAY_LENGTH (direction_strings)))
-    return direction_strings[direction - HB_DIRECTION_LTR];
-
-  return "invalid";
-}
-
-
 /* hb_language_t */
 
 struct hb_language_impl_t {