+++ harfbuzz/src/hb-common.h
@@ -141,13 +141,6 @@ typedef enum {
   HB_DIRECTION_BTT
 } hb_direction_t;
 
-/* len=-1 means str is NUL-terminated */
-HB_EXTERN hb_direction_t
-hb_direction_from_string (const char *str, int len);
-
-HB_EXTERN const char *
-hb_direction_to_string (hb_direction_t direction);
-
 #define HB_DIRECTION_IS_VALID(dir)	((((unsigned int) (dir)) & ~3U) == 4)
 /* Direction must be valid for the following */
 #define HB_DIRECTION_IS_HORIZONTAL(dir)	((((unsigned int) (dir)) & ~1U) == 4)